requirements.txt
streamlit
streamlit-chat
langchain
tiktoken
openai
python-dotenv